backend/app/depth_estimation.py
"""
Depth Estimation and Dimension Measurement for Billboard Detection
Uses monocular depth estimation and camera intrinsics for accurate size measurement
"""

import cv2
import numpy as np
import torch
from PIL import Image
import logging
from typing import Dict, Tuple, Optional
import requests
import os

logger = logging.getLogger(__name__)

class DepthEstimator:
    """Monocular depth estimation for billboard dimension measurement"""
    
    def __init__(self):
        """Initialize depth estimation model"""
        try:
            # In production, would use MiDaS or DPT models
            # For demo, using simplified depth estimation
            self.model_loaded = True
            self.camera_params = self._get_default_camera_params()
            logger.info("Depth estimator initialized")
        except Exception as e:
            logger.error(f"Failed to initialize depth estimator: {e}")
            self.model_loaded = False
    
    def _get_default_camera_params(self) -> Dict:
        """Default camera intrinsic parameters for smartphones"""
        return {
            'focal_length_mm': 26,  # 26mm equivalent
            'sensor_width_mm': 5.76,  # Typical smartphone sensor
            'sensor_height_mm': 4.29,
            'fov_horizontal_deg': 60,
            'fov_vertical_deg': 45
        }
    
    def estimate_depth_map(self, image_path: str) -> np.ndarray:
        """
        Generate depth map from single image
        
        Args:
            image_path: Path to input image
            
        Returns:
            Depth map as numpy array
        """
        try:
            image = cv2.imread(image_path)
            if image is None:
                raise ValueError(f"Could not load image: {image_path}")
            
            # Simplified depth estimation using image gradients and heuristics
            # In production, would use:
            # - MiDaS: https://github.com/isl-org/MiDaS
            # - DPT: https://github.com/isl-org/DPT
            # - ZoeDepth: https://github.com/isl-org/ZoeDepth
            
            gray = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
            
            # Use Sobel gradients as proxy for depth
            grad_x = cv2.Sobel(gray, cv2.CV_64F, 1, 0, ksize=3)
            grad_y = cv2.Sobel(gray, cv2.CV_64F, 0, 1, ksize=3)
            gradient_magnitude = np.sqrt(grad_x**2 + grad_y**2)
            
            # Invert gradients (high gradients = closer objects)
            depth_map = 255 - gradient_magnitude
            depth_map = cv2.GaussianBlur(depth_map, (5, 5), 0)
            
            # Normalize to 0-100 meters range
            depth_map = (depth_map / 255.0) * 100.0
            
            return depth_map.astype(np.float32)
            
        except Exception as e:
            logger.error(f"Depth estimation failed: {e}")
            # Return default depth map
            image = cv2.imread(image_path)
            return np.full(image.shape[:2], 25.0, dtype=np.float32)  # 25m default
    
    def measure_billboard_dimensions(self, image_path: str, bbox: list, gps_coords: Optional[Tuple] = None) -> Dict:
        """
        Measure real-world billboard dimensions using depth estimation
        
        Args:
            image_path: Path to image
            bbox: Bounding box [x1, y1, x2, y2]
            gps_coords: Optional GPS coordinates (lat, lon)
            
        Returns:
            Dict with dimension measurements
        """
        try:
            image = cv2.imread(image_path)
            if image is None:
                raise ValueError(f"Could not load image: {image_path}")
            
            h, w = image.shape[:2]
            x1, y1, x2, y2 = bbox
            
            # Get depth map
            depth_map = self.estimate_depth_map(image_path)
            
            # Extract depth in billboard region
            billboard_region = depth_map[int(y1):int(y2), int(x1):int(x2)]
            avg_depth = np.mean(billboard_region)
            
            # Calculate pixel dimensions
            pixel_width = x2 - x1
            pixel_height = y2 - y1
            
            # Convert to real-world dimensions using camera intrinsics
            real_width = self._pixels_to_meters(pixel_width, avg_depth, w, 'horizontal')
            real_height = self._pixels_to_meters(pixel_height, avg_depth, h, 'vertical')
            
            # Calculate distance from junctions using GPS if available
            junction_distance = None
            if gps_coords:
                junction_distance = self._calculate_junction_distance(gps_coords)
            
            return {
                'dimensions': {
                    'width_meters': float(real_width),
                    'height_meters': float(real_height),
                    'area_sqm': float(real_width * real_height),
                    'depth_meters': float(avg_depth)
                },
                'pixel_dimensions': {
                    'width': float(pixel_width),
                    'height': float(pixel_height)
                },
                'measurement_method': 'monocular_depth_estimation',
                'accuracy': 'medium',  # Would be 'high' with proper depth model
                'junction_distance_meters': junction_distance,
                'camera_params': self.camera_params
            }
            
        except Exception as e:
            logger.error(f"Dimension measurement failed: {e}")
            return {
                'error': str(e),
                'measurement_method': 'failed'
            }
    
    def _pixels_to_meters(self, pixels: float, depth_meters: float, image_dimension: int, direction: str) -> float:
        """
        Convert pixel measurements to real-world meters using camera intrinsics
        
        Args:
            pixels: Pixel measurement
            depth_meters: Distance to object
            image_dimension: Total image width or height in pixels
            direction: 'horizontal' or 'vertical'
            
        Returns:
            Real-world measurement in meters
        """
        if direction == 'horizontal':
            fov_rad = np.radians(self.camera_params['fov_horizontal_deg'])
        else:
            fov_rad = np.radians(self.camera_params['fov_vertical_deg'])
        
        # Calculate field of view at given depth
        fov_at_depth = 2 * depth_meters * np.tan(fov_rad / 2)
        
        # Convert pixels to meters
        meters_per_pixel = fov_at_depth / image_dimension
        return pixels * meters_per_pixel
    
    def _calculate_junction_distance(self, gps_coords: Tuple) -> Optional[float]:
        """
        Calculate distance to nearest traffic junction using GPS
        
        Args:
            gps_coords: (latitude, longitude) tuple
            
        Returns:
            Distance to nearest junction in meters
        """
        try:
            lat, lon = gps_coords
            
            # In production, would query junction database or use:
            # - OpenStreetMap Overpass API for traffic junctions
            # - Google Maps Roads API
            # - Local traffic authority databases
            
            # Mock calculation for demo
            # Assume junctions every ~200m in urban areas
            mock_junctions = [
                (lat + 0.001, lon + 0.001),  # ~100m northeast
                (lat - 0.002, lon + 0.001),  # ~200m southeast
                (lat + 0.001, lon - 0.002),  # ~200m northwest
            ]
            
            min_distance = float('inf')
            for j_lat, j_lon in mock_junctions:
                # Haversine distance calculation
                distance = self._haversine_distance(lat, lon, j_lat, j_lon)
                min_distance = min(min_distance, distance)
            
            return min_distance
            
        except Exception as e:
            logger.error(f"Junction distance calculation failed: {e}")
            return None
    
    def _haversine_distance(self, lat1: float, lon1: float, lat2: float, lon2: float) -> float:
        """Calculate distance between two GPS coordinates using Haversine formula"""
        R = 6371000  # Earth's radius in meters
        
        lat1_rad = np.radians(lat1)
        lat2_rad = np.radians(lat2)
        delta_lat = np.radians(lat2 - lat1)
        delta_lon = np.radians(lon2 - lon1)
        
        a = (np.sin(delta_lat / 2) ** 2 + 
             np.cos(lat1_rad) * np.cos(lat2_rad) * np.sin(delta_lon / 2) ** 2)
        c = 2 * np.arctan2(np.sqrt(a), np.sqrt(1 - a))
        
        return R * c


class ARMarkerMeasurement:
    """Alternative measurement using AR markers for precise calibration"""
    
    def __init__(self):
        """Initialize AR marker detection"""
        self.aruco_dict = cv2.aruco.Dictionary_get(cv2.aruco.DICT_6X6_250)
        self.aruco_params = cv2.aruco.DetectorParameters_create()
        self.marker_size_cm = 10.0  # Standard 10cm AR markers
    
    def detect_and_measure(self, image_path: str, bbox: list) -> Dict:
        """
        Use AR markers for precise billboard measurement
        
        Args:
            image_path: Path to image
            bbox: Billboard bounding box
            
        Returns:
            Measurement results with high accuracy
        """
        try:
            image = cv2.imread(image_path)
            gray = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
            
            # Detect ArUco markers
            corners, ids, _ = cv2.aruco.detectMarkers(gray, self.aruco_dict, parameters=self.aruco_params)
            
            if ids is not None and len(corners) > 0:
                # Use markers for scale calibration
                marker_corners = corners[0][0]  # First marker
                marker_pixel_size = np.linalg.norm(marker_corners[0] - marker_corners[1])
                
                # Calculate scale: pixels per cm
                pixels_per_cm = marker_pixel_size / self.marker_size_cm
                
                # Measure billboard
                x1, y1, x2, y2 = bbox
                pixel_width = x2 - x1
                pixel_height = y2 - y1
                
                real_width_cm = pixel_width / pixels_per_cm
                real_height_cm = pixel_height / pixels_per_cm
                
                return {
                    'dimensions': {
                        'width_meters': real_width_cm / 100.0,
                        'height_meters': real_height_cm / 100.0,
                        'area_sqm': (real_width_cm * real_height_cm) / 10000.0
                    },
                    'measurement_method': 'ar_marker_calibration',
                    'accuracy': 'high',
                    'markers_detected': len(corners),
                    'scale_pixels_per_cm': float(pixels_per_cm)
                }
            else:
                return {
                    'error': 'No AR markers detected',
                    'measurement_method': 'ar_marker_failed'
                }
                
        except Exception as e:
            logger.error(f"AR marker measurement failed: {e}")
            return {
                'error': str(e),
                'measurement_method': 'ar_marker_failed'
            }


def create_measurement_pipeline():
    """Factory function to create measurement pipeline"""
    return {
        'depth_estimator': DepthEstimator(),
        'ar_marker': ARMarkerMeasurement()
    }
