backend/app/detection.py
"""
Billboard Detection Pipeline using YOLOv8
Provides computer vision capabilities for detecting billboards in images
"""

import cv2
import numpy as np
from PIL import Image
import torch
from ultralytics import YOLO
import logging
from typing import Dict, List, Tuple, Optional
import os

logger = logging.getLogger(__name__)

class BillboardDetector:
    """YOLOv8-based billboard detection system"""
    
    def __init__(self, model_path: Optional[str] = None):
        """
        Initialize the billboard detector
        
        Args:
            model_path: Path to custom trained model, defaults to YOLOv8n
        """
        try:
            # Use custom model if provided, otherwise use pre-trained YOLOv8
            if model_path and os.path.exists(model_path):
                self.model = YOLO(model_path)
                logger.info(f"Loaded custom model from {model_path}")
            else:
                # Use YOLOv8n for general object detection
                self.model = YOLO('yolov8n.pt')
                logger.info("Loaded YOLOv8n pre-trained model")
                
            self.confidence_threshold = 0.5
            self.billboard_classes = ['billboard', 'sign', 'advertisement']
            
        except Exception as e:
            logger.error(f"Failed to initialize detector: {e}")
            raise
    
    def detect_billboards(self, image_path: str) -> Dict:
        """
        Detect billboards in an image
        
        Args:
            image_path: Path to the image file
            
        Returns:
            Dict containing detection results
        """
        try:
            # Load and process image
            image = cv2.imread(image_path)
            if image is None:
                raise ValueError(f"Could not load image from {image_path}")
            
            # Run inference
            results = self.model(image, conf=self.confidence_threshold)
            
            detections = []
            for result in results:
                boxes = result.boxes
                if boxes is not None:
                    for box in boxes:
                        # Extract detection data
                        x1, y1, x2, y2 = box.xyxy[0].cpu().numpy()
                        confidence = box.conf[0].cpu().numpy()
                        class_id = int(box.cls[0].cpu().numpy())
                        class_name = self.model.names[class_id]
                        
                        # Filter for billboard-like objects
                        if self._is_billboard_like(class_name, confidence):
                            detection = {
                                'bbox': [float(x1), float(y1), float(x2), float(y2)],
                                'confidence': float(confidence),
                                'class': class_name,
                                'area': float((x2 - x1) * (y2 - y1)),
                                'center': [float((x1 + x2) / 2), float((y1 + y2) / 2)]
                            }
                            detections.append(detection)
            
            return {
                'detections': detections,
                'image_shape': image.shape,
                'detection_count': len(detections),
                'model_info': {
                    'name': 'YOLOv8n',
                    'confidence_threshold': self.confidence_threshold
                }
            }
            
        except Exception as e:
            logger.error(f"Detection failed for {image_path}: {e}")
            return {
                'detections': [],
                'error': str(e),
                'detection_count': 0
            }
    
    def _is_billboard_like(self, class_name: str, confidence: float) -> bool:
        """
        Determine if detected object is billboard-like
        
        Args:
            class_name: YOLO class name
            confidence: Detection confidence
            
        Returns:
            Boolean indicating if object is billboard-like
        """
        # Check for billboard-related classes
        billboard_keywords = ['sign', 'billboard', 'advertisement', 'banner']
        
        if confidence < self.confidence_threshold:
            return False
            
        # Direct match for billboard classes
        if class_name.lower() in self.billboard_classes:
            return True
            
        # Check for billboard-like keywords
        for keyword in billboard_keywords:
            if keyword in class_name.lower():
                return True
                
        return False
    
    def estimate_billboard_size(self, detection: Dict, image_shape: Tuple) -> Dict:
        """
        Estimate billboard physical dimensions using computer vision
        
        Args:
            detection: Detection result from detect_billboards
            image_shape: Shape of the input image
            
        Returns:
            Dict with size estimation results
        """
        try:
            bbox = detection['bbox']
            x1, y1, x2, y2 = bbox
            
            # Calculate pixel dimensions
            pixel_width = x2 - x1
            pixel_height = y2 - y1
            pixel_area = pixel_width * pixel_height
            
            # Estimate real-world dimensions using heuristics
            # This is a simplified approach - in production, would use:
            # - Camera intrinsics and depth estimation
            # - Reference objects for scale
            # - GPS data and known landmark distances
            
            # Assume average billboard viewing distance and camera FOV
            estimated_distance = self._estimate_distance(pixel_area, image_shape)
            estimated_width = self._pixel_to_meters(pixel_width, estimated_distance)
            estimated_height = self._pixel_to_meters(pixel_height, estimated_distance)
            
            return {
                'pixel_dimensions': {
                    'width': float(pixel_width),
                    'height': float(pixel_height),
                    'area': float(pixel_area)
                },
                'estimated_dimensions': {
                    'width_meters': float(estimated_width),
                    'height_meters': float(estimated_height),
                    'area_sqm': float(estimated_width * estimated_height)
                },
                'estimation_method': 'heuristic_distance',
                'confidence': 'medium'
            }
            
        except Exception as e:
            logger.error(f"Size estimation failed: {e}")
            return {
                'error': str(e),
                'estimation_method': 'failed'
            }
    
    def _estimate_distance(self, pixel_area: float, image_shape: Tuple) -> float:
        """Estimate distance to billboard based on pixel area"""
        image_area = image_shape[0] * image_shape[1]
        area_ratio = pixel_area / image_area
        
        # Heuristic: larger objects in frame are typically closer
        # This is a simplified model - real implementation would use:
        # - Monocular depth estimation (MiDaS, DPT)
        # - Stereo vision
        # - LiDAR data if available
        
        if area_ratio > 0.3:
            return 10.0  # Very close
        elif area_ratio > 0.1:
            return 25.0  # Medium distance
        else:
            return 50.0  # Far distance
    
    def _pixel_to_meters(self, pixels: float, distance: float) -> float:
        """Convert pixel measurements to meters using distance estimate"""
        # Simplified conversion using typical camera FOV
        # Real implementation would use camera intrinsics
        fov_horizontal = 60  # degrees
        sensor_width = 0.036  # 36mm (typical smartphone)
        focal_length = 0.026  # 26mm equivalent
        
        # Calculate meters per pixel at given distance
        meters_per_pixel = (distance * sensor_width) / (focal_length * 1000)  # rough approximation
        return pixels * meters_per_pixel * 0.1  # Scale factor for mobile cameras


def create_inference_pipeline():
    """Factory function to create detection pipeline"""
    return BillboardDetector()


# Mock training pipeline documentation
TRAINING_PIPELINE_INFO = {
    "model_architecture": "YOLOv8n",
    "training_data": {
        "dataset": "Custom billboard dataset + COCO signs",
        "images": "5000+ annotated billboard images",
        "classes": ["billboard", "sign", "advertisement", "banner"],
        "augmentations": ["rotation", "scaling", "color_jitter", "blur"]
    },
    "training_config": {
        "epochs": 100,
        "batch_size": 16,
        "learning_rate": 0.001,
        "optimizer": "AdamW",
        "image_size": 640
    },
    "performance_metrics": {
        "mAP@0.5": 0.78,
        "precision": 0.82,
        "recall": 0.75,
        "inference_time": "45ms on GPU"
    },
    "model_weights": {
        "location": "models/billboard_yolov8_best.pt",
        "size": "6.2MB",
        "download_url": "https://github.com/jayden-crypto/hackathon-billboard-sentinel/releases/download/v1.0/billboard_yolov8_best.pt"
    }
}
